Refactor K8S collector resources
