+++ plugins/k8s/cloudkeeper_plugin_k8s/resources/__init__.py
@@ -1,0 +1,3 @@
+from .cluster import KubernetesCluster
+from .namespace import KubernetesNamespace
+from .pod import KubernetesPod
