+++ plugins/k8s/cloudkeeper_plugin_k8s/resources/namespace.py
@@ -1,0 +1,10 @@
+import cloudkeeper.logging
+from .common import KubernetesResource
+from cloudkeeper.baseresources import BaseRegion
+
+
+log = cloudkeeper.logging.getLogger("cloudkeeper." + __name__)
+
+
+class KubernetesNamespace(KubernetesResource, BaseRegion):
+    resource_type = "kubernetes_namespace"
